build: copier-auto-update
